+++ apps/rig_auth/lib/rig_auth/authorization_check/subscription.ex
@@ -4,9 +4,9 @@ defmodule RigAuth.AuthorizationCheck.Subscription do
   """
   use Rig.Config, :custom_validation
 
-  alias Plug.Conn
   alias RigAuth.AuthorizationCheck.External
   alias RigAuth.AuthorizationCheck.Header
+  alias RigAuth.AuthorizationCheck.Request
 
   # Confex callback
   defp validate_config!(config) do
@@ -26,8 +26,15 @@ defmodule RigAuth.AuthorizationCheck.Subscription do
     }
   end
 
-  @spec check_authorization(Conn.t()) :: :ok | {:error, :not_authorized}
-  def check_authorization(conn) do
+  # ---
+
+  @spec check_authorization(Request.t()) :: :ok | {:error, :not_authorized}
+  def check_authorization(request)
+
+  # If body is nil, there are no subscriptions to authorize.
+  def check_authorization(%{body: nil}), do: :ok
+
+  def check_authorization(request) do
     %{validation_type: validation_type} = config()
 
     case validation_type do
@@ -35,14 +42,14 @@ defmodule RigAuth.AuthorizationCheck.Subscription do
         :ok
 
       :jwt_validation ->
-        if Header.any_valid_bearer_token?(conn) do
+        if Header.any_valid_bearer_token?(request) do
           :ok
         else
           {:error, :not_authorized}
         end
 
       {:url, base_url} ->
-        External.check_or_log(base_url, conn.req_headers, conn.body_params)
+        External.check_or_log(base_url, request)
     end
   end
 end