+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [Proxy] Fixed Kafka request logger in proxy -- now supports also Apache Avro. [#170](https://github.com/Accenture/reactive-interaction-gateway/issues/170)
+- Fixed usage of external check for `SUBMISSION_CHECK` and `SUBSCRIPTION_CHECK`. [#241](https://github.com/Accenture/reactive-interaction-gateway/issues/241)
 
 <!-- ### Deprecated -->